+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,10 +3,9 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.HandlerAdapterRegistry;
-import nextstep.mvc.controller.tobe.HandlerExecutionHandlerAdapter;
-import nextstep.mvc.controller.asis.ControllerHandlerAdaptor;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
 import nextstep.mvc.HandlerMappingRegistry;
+import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.tobe.HandlerExecutionHandlerAdapter;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,10 +20,8 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
         final HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
 
         final var dispatcherServlet = new DispatcherServlet(handlerMappingRegistry, handlerAdapterRegistry);
-        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
         dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller"));
         dispatcherServlet.addHandlerAdapter(new HandlerExecutionHandlerAdapter());
-        dispatcherServlet.addHandlerAdapter(new ControllerHandlerAdaptor());
 
         final var dispatcher = servletContext.addServlet("dispatcher", dispatcherServlet);
         dispatcher.setLoadOnStartup(1);
