+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -7,10 +7,10 @@ import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
 
 @Controller
-public class RegisterViewController {
+public class ForwardController {
 
-    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) {
-        return "/register.jsp";
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public String execute(final HttpServletRequest request, final HttpServletResponse response) {
+        return "/index.jsp";
     }
 }
