--- mvc/src/main/java/nextstep/mvc/controller/asis/ForwardController.java
@@ -1,19 +1,0 @@
-package nextstep.mvc.controller.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.Objects;
-
-public class ForwardController implements Controller {
-
-    private final String path;
-
-    public ForwardController(final String path) {
-        this.path = Objects.requireNonNull(path);
-    }
-
-    @Override
-    public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return path;
-    }
-}
